Revamp/public txn history (#1028)
#994 #818 

This PR adds:

- Adds a Medallion-style public history pipeline: Bronze ingestion,
Silver transfer-leg normalization, and Gold frontend-ready history
projection.
- Uses Goldsky as the activity trigger to enqueue targeted RefreshLatest
jobs, reducing reliance on polling and making new treasury activity
visible faster.
- Adds proposal receipt linking and a periodic reconciler to safely
converge proposal statuses missed during receipt processing.
- Adds an in-process SSE event bus and /events endpoint so the frontend
can refetch treasury data when projections change instead of polling
(can be use at other places to avoid polling and get fresh data).
- Adds token price ingestion with canonical token registry, 5-minute USD
price series, and monthly Postgres range partitions.
- Adds shared Postgres-backed Apalis job queue infrastructure with
deduped in-flight jobs and dedicated public history workers for
RefreshLatest and BackfillPage.

+++ nt-be/tests/confidential_monitor_test.rs
@@ -200,11 +200,7 @@ async fn handle_confidential_outgoing_external_payment(pool: PgPool) {
 
     let (amount, counterparty, token_id, method_name, raw_data) = row;
     assert_eq!(token_id, "intents.near:nep141:wrap.near");
-    assert!(
-        amount < BigDecimal::from(0),
-        "outgoing leg must be negative, got {}",
-        amount
-    );
+    assert!(amount < 0, "outgoing leg must be negative, got {}", amount);
     assert_eq!(
         counterparty, EXTERNAL_RECIPIENT,
         "payment counterparty should be the quote recipient"
